Revert "setup refactor for projectlayout"
This reverts commit 0c309338983d3cc71cb1ec46ee9bb12e1363422a.
