+++ submit-web/src/components/Projects/Project.tsx
@@ -1,14 +1,16 @@
-import { Box, Button, Paper, styled, Typography } from "@mui/material";
+import { Box, Button, Divider, Paper, styled, Typography } from "@mui/material";
 import { BCDesignTokens } from "epic.theme";
 import AddIcon from "@mui/icons-material/Add";
 import { ProjectStatus } from "../registration/addProjects/ProjectStatus";
 import { PROJECT_STATUS } from "../registration/addProjects/ProjectCard/constants";
+import SubmissionPackageTable from "./ProjectTable";
 import { AccountProject } from "@/models/Project";
-import { Outlet, useNavigate } from "@tanstack/react-router";
+import { PACKAGE_STATUS } from "@/models/Package";
+import { useNavigate } from "@tanstack/react-router";
 
 const HEADER_HEIGHT = 54;
 
-export const CardInnerBox = styled(Box)({
+const CardInnerBox = styled(Box)({
   display: "flex",
   alignItems: "flex-start",
   justifyContent: "center",
@@ -23,6 +25,13 @@ type ProjectParam = {
 
 export const Project = ({ accountProject }: ProjectParam) => {
   const navigate = useNavigate();
+
+  const activeSubmissionPackages = accountProject.packages.filter(
+    (subPackage) => subPackage.status === PACKAGE_STATUS.IN_REVIEW.value,
+  );
+  const pastSubmissionPackages = accountProject.packages.filter(
+    (subPackage) => subPackage.status !== PACKAGE_STATUS.IN_REVIEW.value,
+  );
 
   const { name, ea_certificate } = accountProject.project;
 
@@ -100,7 +109,48 @@ export const Project = ({ accountProject }: ProjectParam) => {
               </Button>
             </CardInnerBox>
           </Box>
-          <Outlet />
+          <Box height={"100%"} px={BCDesignTokens.layoutPaddingXsmall}>
+            <Divider sx={{ mb: 0.5 }} />
+            <Typography
+              variant="body1"
+              sx={{
+                fontWeight: "bold",
+                backgroundColor: BCDesignTokens.themeGold10,
+              }}
+            >
+              Active Submissions
+            </Typography>
+            <CardInnerBox
+              sx={{ height: "100%", py: BCDesignTokens.layoutPaddingSmall }}
+            >
+              <SubmissionPackageTable
+                submissionPackages={activeSubmissionPackages}
+              />
+            </CardInnerBox>
+            <Divider
+              sx={{
+                mb: BCDesignTokens.layoutPaddingXsmall,
+                mt: BCDesignTokens.layoutPaddingSmall,
+              }}
+            />
+            <Typography
+              variant="body1"
+              sx={{
+                fontWeight: "bold",
+                backgroundColor: BCDesignTokens.themeGold10,
+              }}
+            >
+              Past Submissions
+            </Typography>
+            <CardInnerBox
+              sx={{ height: "100%", py: BCDesignTokens.layoutPaddingMedium }}
+            >
+              <SubmissionPackageTable
+                headless
+                submissionPackages={pastSubmissionPackages}
+              />
+            </CardInnerBox>
+          </Box>
         </Box>
       </Box>
     </Paper>
