+++ submit-web/src/components/Projects/ProjectTable.tsx
@@ -17,11 +17,9 @@ import { SubmissionPackage } from "@/models/Package";
 export default function SubmissionPackageTable({
   submissionPackages,
   headless,
-  onSubmissionClick,
 }: {
   submissionPackages: Array<SubmissionPackage>;
   headless?: boolean;
-  onSubmissionClick: (submissionId: number) => void;
 }) {
   return (
     <TableContainer component={Box} sx={{ height: "100%" }}>
@@ -88,8 +86,6 @@ export default function SubmissionPackageTable({
                     display: "flex",
                     alignItems: "center",
                   }}
-                  component={"button"}
-                  onClick={() => onSubmissionClick(subPackage.id)}
                 >
                   <Typography
                     variant="h5"
